backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.7.3
scikit-learn==1.3.2
pandas==2.2.2
numpy==1.26.4
redis==5.3.1
httpx==0.24.1
tenacity==8.2.2
